requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dash==2.1.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jws==0.1.3
MarkupSafe==2.0.1
networkx==2.6.3
numpy==1.22.2
oauth2client==4.1.3
pandas==1.4.0
plotly==5.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.0
pyparsing==3.0.7
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
tenacity==8.0.1
urllib3==1.26.8
Werkzeug==2.0.2
